CCIDIST-28: We now support issue list paging properly.
We only load the requested range from the DB. Some code to get prev/next
link has also been simplified.

+++ ccidist-persistence/src/main/java/com/cefalo/cci/dao/IssueDao.java
@@ -10,7 +10,7 @@ public interface IssueDao {
 
     List<Issue> getIssueListByPublicationId(String publicationId);
 
-    List<Issue> getIssueListByPublicationId(String publicationId, int start, int maxResult);
+    List<Issue> getIssueListByPublicationId(String publicationId, long start, long maxResult);
 
     EpubFile getEpubFile(long id);
